+++ core/tracing/journal.go
@@ -38,9 +38,8 @@ type revision struct {
 // journal is a state change journal to be wrapped around a tracer.
 // It will emit the state change hooks with reverse values when a call reverts.
 type journal struct {
-	entries     []entry
-	hooks       *Hooks
-	lastCreator *common.Address // Account that initiated the last contract creation
+	entries []entry
+	hooks   *Hooks
 
 	validRevisions []revision
 	nextRevisionId int
@@ -57,8 +56,11 @@ func WrapWithJournal(hooks *Hooks) (*Hooks, error) {
 		return nil, fmt.Errorf("wrapping nil tracer")
 	}
 	// No state change to journal, return the wrapped hooks as is
-	if hooks.OnBalanceChange == nil && hooks.OnNonceChange == nil && hooks.OnCodeChange == nil && hooks.OnStorageChange == nil {
+	if hooks.OnBalanceChange == nil && hooks.OnNonceChange == nil && hooks.OnNonceChangeV2 == nil && hooks.OnCodeChange == nil && hooks.OnStorageChange == nil {
 		return hooks, nil
+	}
+	if hooks.OnNonceChange != nil && hooks.OnNonceChangeV2 != nil {
+		return nil, fmt.Errorf("cannot have both OnNonceChange and OnNonceChangeV2")
 	}
 
 	// Create a new Hooks instance and copy all hooks
@@ -73,8 +75,12 @@ func WrapWithJournal(hooks *Hooks) (*Hooks, error) {
 	if hooks.OnBalanceChange != nil {
 		wrapped.OnBalanceChange = j.OnBalanceChange
 	}
-	if hooks.OnNonceChange != nil {
-		wrapped.OnNonceChange = j.OnNonceChange
+	if hooks.OnNonceChange != nil || hooks.OnNonceChangeV2 != nil {
+		// Regardless of which hook version is used in the tracer,
+		// the journal will want to capture the nonce change reason.
+		wrapped.OnNonceChangeV2 = j.OnNonceChangeV2
+		// A precaution to ensure EVM doesn't call both hooks.
+		wrapped.OnNonceChange = nil
 	}
 	if hooks.OnCodeChange != nil {
 		wrapped.OnCodeChange = j.OnCodeChange
@@ -142,18 +148,12 @@ func (j *journal) OnTxEnd(receipt *types.Receipt, err error) {
 
 func (j *journal) OnEnter(depth int, typ byte, from common.Address, to common.Address, input []byte, gas uint64, value *big.Int) {
 	j.revIds = append(j.revIds, j.snapshot())
-	if typ == CREATE || typ == CREATE2 {
-		j.lastCreator = &from
-	}
 	if j.hooks.OnEnter != nil {
 		j.hooks.OnEnter(depth, typ, from, to, input, gas, value)
 	}
 }
 
 func (j *journal) OnExit(depth int, output []byte, gasUsed uint64, err error, reverted bool) {
-	if j.lastCreator != nil {
-		j.lastCreator = nil
-	}
 	revId := j.revIds[len(j.revIds)-1]
 	j.revIds = j.revIds[:len(j.revIds)-1]
 	if reverted {
@@ -171,16 +171,15 @@ func (j *journal) OnBalanceChange(addr common.Address, prev, new *big.Int, reaso
 	}
 }
 
-func (j *journal) OnNonceChange(addr common.Address, prev, new uint64) {
+func (j *journal) OnNonceChangeV2(addr common.Address, prev, new uint64, reason NonceChangeReason) {
 	// When a contract is created, the nonce of the creator is incremented.
 	// This change is not reverted when the creation fails.
-	if j.lastCreator != nil && *j.lastCreator == addr {
-		// Skip only the first nonce change.
-		j.lastCreator = nil
-	} else {
+	if reason != NonceChangeContractCreator {
 		j.entries = append(j.entries, nonceChange{addr: addr, prev: prev, new: new})
 	}
-	if j.hooks.OnNonceChange != nil {
+	if j.hooks.OnNonceChangeV2 != nil {
+		j.hooks.OnNonceChangeV2(addr, prev, new, reason)
+	} else if j.hooks.OnNonceChange != nil {
 		j.hooks.OnNonceChange(addr, prev, new)
 	}
 }
@@ -241,7 +240,9 @@ func (b balanceChange) revert(hooks *Hooks) {
 }
 
 func (n nonceChange) revert(hooks *Hooks) {
-	if hooks.OnNonceChange != nil {
+	if hooks.OnNonceChangeV2 != nil {
+		hooks.OnNonceChangeV2(n.addr, n.new, n.prev, NonceChangeRevert)
+	} else if hooks.OnNonceChange != nil {
 		hooks.OnNonceChange(n.addr, n.new, n.prev)
 	}
 }